+++ src/main/java/net/paladins/config/TweaksConfig.java
@@ -1,6 +1,7 @@
 package net.paladins.config;
 
-public class EffectsConfig {
+public class TweaksConfig {
+    public boolean ignore_items_required_mods = false;
     public float battle_banner_attack_speed_bonus = 0.4F;
     public float battle_banner_ranged_haste_bonus = 0.4F;
     public float battle_banner_spell_haste_bonus = 0.4F;